Restore original meaning of run_stats
The run_stats config flag tells which instance should update
the counters generally.

This commit refactors the queue stats handling so that the
rx/tx counters are always updated regardless of whether
run_stats is used or not. Only these counters are updated
unless the app is the run_stats instance.

Also avoid updating the link_status, etc counters in each
app since those are global to the NIC